Switch HoT package to Base type
