+++ wrappers/vue/.config/minified.js
@@ -14,7 +14,7 @@ const minConfig = {
     exports: 'named',
     globals: {
       vue: 'Vue',
-      handsontable: 'Handsontable'
+      'handsontable/base': 'Handsontable'
     }
   },
   plugins: baseConfig.plugins.concat([
